requirements.txt
vllm
openai
aiohttp
tqdm
pandas
pyarrow
pyyaml
scikit-learn
pydantic
wandb
sentence-transformers

jinja2
